More default error handling for when _basePlayer doesn't exist
Primary motivation was when  Widevine fails to initialize, we actually do NOT have a baseMoviePlayer initialized.  However, the UIs are all initialized and try to update with basePlayer data

We need to make sure that no Getters null-point, and that Error/State related Getters/Setters store their info on the MoviePlayer.

Most setters and void methods should not be called if the player does not exist, and don't need to be wrapped with the test