[PBK-274] Add more Auth errors.
None of the Device Management error auth codes were handled properly

Also, there's no reason to replace the auth error with -2.  Why would we need to do that?  The only possible use is when indexing into our list of authorization descriptions, but when we do that we check valid indexing.